+++ gapi_plus/RGenerator.cs
@@ -19,7 +19,7 @@ namespace gapi_plus {
 
 			var acts = Service.Activities.List(setting.ProfileID, setting.Collection);
 
-			var step = setting.MaxResults / 100;
+			var step = setting.MaxResults / (double)100;
 			string nextPage = string.Empty;
 
 			var users = new Dictionary<string, bool>();
